chore: Move useElement hook into its own file
